--- tests/convergence/conv_rates.hpp
@@ -1,77 +1,0 @@
-
-#include<mfem.hpp>
-#include <fstream>
-#include <iostream>
-
-using namespace std;
-using namespace mfem;
-
-class Convergence
-{
-private:
-
-#ifdef MFEM_USE_MPI
-   MPI_Comm comm;
-#endif
-   int counter=0;
-   int dcounter=0;
-   int fcounter=0;
-   int cont_type;
-   int print_flag=1;
-   Array<double> L2Errors;
-   Array<double> DGFaceErrors;
-   Array<double> DErrors;
-   Array<double> EnErrors;
-   Array<double> DGFaceRates;
-   Array<double> L2Rates;
-   Array<double> DRates;
-   Array<double> EnRates;
-   Array<int> ndofs;
-   double CoeffNorm;
-   double CoeffDNorm;
-   void AddL2Error(GridFunction * gf, Coefficient * u, VectorCoefficient * U);
-   void AddGf(GridFunction * gf, Coefficient * u, VectorCoefficient * grad,
-              Coefficient * ell_coeff, double Nu);
-   void AddGf(GridFunction * gf, VectorCoefficient * u, 
-              VectorCoefficient * curl, Coefficient * div);
-
-   double GetNorm(GridFunction * gf, Coefficient * u, VectorCoefficient * U);
-public:
-   Convergence();
-#ifdef MFEM_USE_MPI
-   Convergence(MPI_Comm _comm);
-#endif
-   // Clear any internal data
-   void Clear();
-
-   void AddGridFunction(GridFunction * gf, Coefficient * u, VectorCoefficient * grad=NULL,
-                        Coefficient * ell_coeff=NULL, double Nu=1.0)
-   {      
-      AddGf(gf, u, grad, ell_coeff,Nu);
-   }
-   void AddGridFunction(GridFunction * gf, VectorCoefficient * u)
-   {
-      AddGf(gf,u, nullptr, nullptr);
-   }
-   void AddGridFunction(GridFunction * gf, VectorCoefficient * u, VectorCoefficient * curl)
-   {
-      AddGf(gf,u, curl, nullptr);
-   }
-   void AddGridFunction(GridFunction * gf, VectorCoefficient * u, Coefficient * div) 
-   {
-      AddGf(gf,u, nullptr, div);
-   }
-   // Get L2 error for step n
-   double GetL2Error(int n);
-
-   // Get all L2 errors
-   Array<double> * GetL2Errors();
-
-   // Get rates for all refinements
-   Array<double> * GetRates();
-
-   // Print rates and errors
-   void Print(bool relative = false);
-
-   ~Convergence(){};
-};
